fix: address Cubic's PR feedback on dependency injection
Fixes identified issues while maintaining PR objective of removing
@calcom/web dependencies from packages via dependency injection:

1. Code Duplication (Comment #2):
   - Extracted buildLegacyRequest to packages/lib/legacy-request.ts
   - Uses generic interfaces (HeadersLike, CookiesLike) instead of Next.js types
   - Both apps/web and packages can now import from shared location
   - Removed unused NextApiRequest import from calcomHandler.ts

2. Missing Cache Revalidation (Comments #3, #5, #8):
   - Added onInvalidate prop to NewWebhookView and wired from webhooks/new page
   - Added onInvalidate prop to EventWebhooksTab (needs wiring from apps/web)
   - Added onInvalidate prop to DuplicateDialog (needs wiring from apps/web)
   - Pattern: server actions with revalidatePath at call sites

3. Promise Handling Bug (Comments #6, #12):
   - Fixed MemberInvitationModal clipboard promise to handle rejections
   - Added reject handler to prevent promise hanging on mutateAsync failure

Note: EventWebhooksTab and DuplicateDialog onInvalidate props are added
but not yet wired from apps/web pages. This requires threading through
multiple component layers and will be completed in follow-up commits.

Platform DI changes (UserDropdown, PlatformMembersView, InviteMemberModal)
were reverted in previous commit to prevent regressions.
