+++ packages/features/auth/signup/handlers/calcomHandler.ts
@@ -1,6 +1,5 @@
 import { cookies, headers } from "next/headers";
 import { NextResponse } from "next/server";
-import type { NextApiRequest } from "next";
 
 import { getPremiumMonthlyPlanPriceId } from "@calcom/app-store/stripepayment/lib/utils";
 import { getLocaleFromRequest } from "@calcom/features/auth/lib/getLocaleFromRequest";
@@ -14,6 +13,7 @@ import { checkIfEmailIsBlockedInWatchlistController } from "@calcom/features/wat
 import { hashPassword } from "@calcom/lib/auth/hashPassword";
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { HttpError } from "@calcom/lib/http-error";
+import { buildLegacyRequest } from "@calcom/lib/legacy-request";
 import logger from "@calcom/lib/logger";
 import type { CustomNextApiHandler } from "@calcom/lib/server/username";
 import { usernameHandler } from "@calcom/lib/server/username";
@@ -30,33 +30,6 @@ import {
 } from "../utils/token";
 
 const log = logger.getSubLogger({ prefix: ["signupCalcomHandler"] });
-
-type HeadersLike = { entries(): IterableIterator<[string, string]> };
-type CookiesLike = { getAll(): { name: string; value: string }[] };
-
-const createProxifiedObject = (object: Record<string, string>) =>
-  new Proxy(object, {
-    set: () => {
-      throw new Error("You are trying to modify 'headers' or 'cookies', which is not supported in app dir");
-    },
-  });
-
-const buildLegacyHeaders = (headers: HeadersLike) => {
-  const headersObject = Object.fromEntries(headers.entries());
-  return createProxifiedObject(headersObject);
-};
-
-const buildLegacyCookies = (cookies: CookiesLike) => {
-  const cookiesObject = cookies.getAll().reduce<Record<string, string>>((acc, { name, value }) => {
-    acc[name] = value;
-    return acc;
-  }, {});
-  return createProxifiedObject(cookiesObject);
-};
-
-const buildLegacyRequest = (headers: HeadersLike, cookies: CookiesLike) => {
-  return { headers: buildLegacyHeaders(headers), cookies: buildLegacyCookies(cookies) } as NextApiRequest;
-};
 
 const handler: CustomNextApiHandler = async (body, usernameStatus) => {
   const {
